requirements.txt
django==3.1.7
requests==2.25.1
djangorestframework==3.12.2
django-cors-headers==3.7.0